getProductListUseCase 작성

+++ app/src/main/java/com/example/aop_part5_chapter02/data/response/ProductResponse.kt
@@ -17,7 +17,7 @@ data class ProductResponse(
             id = id.toLong(),
             createdAt = Date(createdAt),
             product_name = product_name,
-            productPrice = productPrice.toInt(),
+            productPrice = productPrice.toDouble().toInt(),
             productImage = productImage,
             productType = productType,
             productIntroductionImage = productIntroductionImage
